+++ agent/api/ecsclient/client_test.go
@@ -675,7 +675,6 @@ func TestSubmitTaskStateChangeWithAttachments(t *testing.T) {
 	assert.NoError(t, err, "Unable to submit task state change with attachments")
 }
 
-//
 func TestSubmitTaskStateChangeWithoutAttachments(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
@@ -695,4 +694,61 @@ func TestSubmitTaskStateChangeWithoutAttachments(t *testing.T) {
 		Status:  api.TaskRunning,
 	})
 	assert.NoError(t, err, "Unable to submit task state change with no attachments")
+}
+
+// TestSubmitContainerStateChangeWhileTaskInPending tests the container state change was submitted
+// when the task is still in pending state
+func TestSubmitContainerStateChangeWhileTaskInPending(t *testing.T) {
+	mockCtrl := gomock.NewController(t)
+	defer mockCtrl.Finish()
+
+	testCases := []struct {
+		taskStatus api.TaskStatus
+	}{
+		{
+			api.TaskStatusNone,
+		},
+		{
+			api.TaskPulled,
+		},
+		{
+			api.TaskCreated,
+		},
+	}
+
+	taskStateChangePending := api.TaskStateChange{
+		Status:  api.TaskCreated,
+		TaskARN: "arn",
+		Containers: []api.ContainerStateChange{
+			{
+				TaskArn:       "arn",
+				ContainerName: "container",
+				Status:        api.ContainerRunning,
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(fmt.Sprintf("TaskStatus: %s", tc.taskStatus.String()), func(t *testing.T) {
+			taskStateChangePending.Status = tc.taskStatus
+			client, _, mockSubmitStateClient := NewMockClient(mockCtrl, ec2.NewBlackholeEC2MetadataClient(), nil)
+			mockSubmitStateClient.EXPECT().SubmitTaskStateChange(&taskSubmitInputMatcher{
+				ecs.SubmitTaskStateChangeInput{
+					Cluster: strptr(configuredCluster),
+					Task:    strptr("arn"),
+					Status:  strptr("PENDING"),
+					Reason:  strptr(""),
+					Containers: []*ecs.ContainerStateChange{
+						{
+							ContainerName:   strptr("container"),
+							Status:          strptr("RUNNING"),
+							NetworkBindings: []*ecs.NetworkBinding{},
+						},
+					},
+				},
+			})
+			err := client.SubmitTaskStateChange(taskStateChangePending)
+			assert.NoError(t, err)
+		})
+	}
 }